+++ launchpy/__init__.py
@@ -6,3 +6,4 @@ May  15 12:04:49 2019
 from .__version__ import __version__
 from .launchpy import *
 from .configs import *
+from .library import *