+++ launchpy/launchpy.py
@@ -1,4 +1,3 @@
-import time
 import json
 from collections import defaultdict
 from concurrent import futures
@@ -9,8 +8,8 @@ import datetime
 import pandas as pd
 from pathlib import Path
 from launchpy import config, connector
-
-# Set up default values
+from typing import IO, Union
+from .library import Library
 
 
 def saveFile(data:str,filename:str=None,type:str='txt',encoding:str='utf-8')->None:
@@ -206,6 +205,28 @@ class Admin:
             return property_class
         else:
             return new_property['data']
+    
+    def updateProperty(self,propertyId:str=None,attributes:dict=None)->dict:
+        """
+        Update a property based on the information passed in the attributes.
+        Arguments:
+            propertyId : REQUIRED : The property ID to be updated
+            attributes : REQUIRED : the dictionary containing the attributes to be updated
+        """
+        if propertyId is None:
+            raise ValueError('Require a property ID')
+        if attributes is None:
+            raise ValueError('require a dictionary with the data')
+        attributes = {**attributes}
+        obj = {"data": {
+                "attributes": attributes,
+                "id": propertyId,
+                "type": "properties"
+            }
+        }
+        path = f"/properties/{propertyId}"
+        res:dict = self.connector.patchData(self.endpoint+path,data=obj)
+        return res
 
 
     def getExtensionsCatalogue(self, availability: str = None, name: str = None, platform: str = "web", save: bool = False)->list:
@@ -581,6 +602,18 @@ class Property:
             else:
                 expanded_list.append(element)
         return expanded_list
+    
+    def getRuleComponent(self,rc_id:str=None)->dict:
+        """
+        Return a ruleComponent information
+        Argument:
+            rc_id : REQUIRED : Rule Component ID
+        """
+        if rc_id is None:
+            raise ValueError('Require a ruleComponent ID')
+        path = f"/rule_components/{rc_id}"
+        res:dict = self.connector.getData(self.endpoint+path)
+        return res
 
     def getDataElements(self)->object:
         """
@@ -1065,7 +1098,7 @@ class Property:
             data = rules
         return data
 
-    def updateRuleComponents(self, rc_id: str, attr_dict: object, **kwargs)->object:
+    def updateRuleComponent(self, rc_id: str, attr_dict: object, **kwargs)->object:
         """
         Update the ruleComponents based on the information provided.
         arguments: 
@@ -1086,6 +1119,50 @@ class Property:
         except:
             data = rc
         return data
+    
+    def updateCustomCode(self,rc_id:str=None,customCode:Union[str,IO]=None,encoding:str='utf-16')->dict:
+        """
+        Update the custom code of a rule (analytics or core)
+        Arguments:
+            rc_id : REQUIRED : Rule Component ID
+            customCode : REQUIRED : code to be updated in the ruleComponent.2 options:
+                javaScript file; example : "myCode.js" -> ".js" suffix is important
+                string; the code you want to write as a string.
+            encoding: OPTIONAL : encoding to read the JS file. Default (utf-16)
+        """
+        if rc_id is None:
+            raise ValueError('Require a ruleComponent ID')
+        if customCode is None:
+            raise ValueError('Require some code to update')
+        if '.js' in customCode:
+            with open(customCode,'r',encoding=encoding) as f:
+                myCode = f.read()
+        else:
+            myCode = customCode
+        myRC = self.getRuleComponent(rc_id=rc_id)
+        myRCsettings = json.loads(myRC['data']['attributes']['settings'])
+        if 'source' in myRCsettings.keys():
+            myRCsettings['source'] = myCode
+        elif 'customSetup' in myRCsettings.keys():
+            myRCsettings['customSetup']['source'] = myCode
+        myNewSettings = json.dumps(myRCsettings)
+        obj = {
+            "data": {
+                "attributes" :{
+                    "settings": myNewSettings
+                },
+                "type": "rule_components",
+                "id": rc_id
+            }
+        }
+        path = f'/rule_components/{rc_id}'
+        rc = self.connector.patchData(self.endpoint+path, data=obj)
+        try:
+            data = rc['data']
+        except:
+            data = rc
+        return data
+
 
     def updateDataElements(self, dataElement_id: str, attr_dict: object, **kwargs)->object:
         """
@@ -1207,95 +1284,6 @@ class Property:
         data = self.connector.deleteData(
             'https://reactor.adobe.io/environments/'+env_id)
         return data
-
-    # Not supported for the moment
-    # def extractAnalyticsConfig(self)->object:
-    #     """
-    #     Extract the analytics configuration that has been done in the Analytics Extensions and Rules.
-    #     Return a dictionary of the different element in a dataframe
-    #     """
-    #     dict_eVars = defaultdict(list)
-    #     dict_props = defaultdict(list)
-    #     dict_events = defaultdict(list)
-    #     dict_value_eVars = defaultdict(list)
-    #     dict_value_props = defaultdict(list)
-    #     p_rules = self.getRules()
-    #     p_ext = self.getExtensions()
-    #     p_rcs = self.getRuleComponents()
-    #     analytics = [ext for ext in p_ext if ext['attributes']
-    #                  ['name'] == 'adobe-analytics'][0]
-    #     analytics_rcs = [rc for rc in p_rcs if rc['attributes']['delegate_descriptor_id'].find(
-    #         'adobe-analytics::actions::set-variables') - 1]
-
-    #     def searchSetupAnalytics(element: object, verbose: bool = False):
-    #         """
-    #         fills the different dictionaries with where informations are held.
-    #         """
-    #         if element['type'] == "rule_components":
-    #             name = element['rule_name']
-    #         elif element['type'] == "extensions":
-    #             name = 'Analytics Extension'
-    #         settings = json.loads(element['attributes']['settings'])
-    #         if 'trackerProperties' in settings.keys():
-    #             tracker_properties = settings['trackerProperties']
-    #         else:
-    #             tracker_properties = {}
-    #         if verbose:
-    #             print(name)
-    #         if len(tracker_properties) > 0:
-    #             if 'eVars' in tracker_properties.keys():
-    #                 for v in tracker_properties['eVars']:
-    #                     dict_eVars[v['name']].append(f'{name} - Interface')
-    #                     dict_value_eVars[v['name']].append(v['value'])
-    #             if 'props' in tracker_properties.keys():
-    #                 for p in tracker_properties['props']:
-    #                     dict_props[p['name']].append(f'{name} - Interface')
-    #                     dict_value_props[p['name']].append(p['value'])
-    #             if 'events' in tracker_properties.keys():
-    #                 for e in tracker_properties['events']:
-    #                     dict_events[e['name']].append(f'{name} - Interface')
-    #         if 'customSetup' in settings.keys():
-    #             code = settings['customSetup']['source']
-    #             if len(code) > 0:
-    #                 matchevents = re.findall('(event[0-9]+)', code)
-    #                 matcheVars = re.findall('(eVar[0-9]+)\s*=', code)
-    #                 matchprops = re.findall('(prop[0-9]+?)\s*=', code)
-    #                 if matcheVars is not None:
-    #                     for v in set(matcheVars):
-    #                         value = f'{name} - Custom Code'
-    #                         if value not in dict_eVars[v]:
-    #                             dict_eVars[v].append(f'{name} - Custom Code')
-    #                 if matchprops is not None:
-    #                     for p in set(matchprops):
-    #                         value = f'{name} - Custom Code'
-    #                         if value not in dict_props[p]:
-    #                             dict_props[p].append(f'{name} - Custom Code')
-    #                 if matchevents is not None:
-    #                     for e in set(matchevents):
-    #                         value = f'{name} - Custom Code'
-    #                         if value not in dict_events[e]:
-    #                             dict_events[e].append(f'{name} - Custom Code')
-    #     searchSetupAnalytics(analytics)
-    #     for rc in analytics_rcs:
-    #         searchSetupAnalytics(rc)
-    #     df_eVars = pd.DataFrame(
-    #         dict([(k, pd.Series(v)) for k, v in dict_eVars.items()])).T.fillna('')
-    #     df_eVars.columns = ['location ' +
-    #                         str(i) for i in range(1, len(df_eVars.columns)+1)]
-    #     df_props = pd.DataFrame(
-    #         dict([(k, pd.Series(v)) for k, v in dict_props.items()])).T.fillna('')
-    #     df_props.columns = ['location ' +
-    #                         str(i) for i in range(1, len(df_props.columns)+1)]
-    #     df_events = pd.DataFrame(
-    #         dict([(k, pd.Series(v)) for k, v in dict_events.items()])).T.fillna('')
-    #     df_events.columns = ['location ' +
-    #                          str(i) for i in range(1, len(df_events.columns)+1)]
-    #     data = {'eVars': df_eVars, 'props': df_props, 'events': df_events}
-    #     return data
-
-
-
-
 
 def extensionsInfo(data: list)->dict:
     """
@@ -1730,307 +1718,3 @@ def extractAnalyticsCode(rcSettings: str, save: bool = False, filename: str = No
         if save:
             saveFile(json_code,filename,type='js',encoding='utf-16')
         return json_code
-
-
-class Library:
-
-    def __init__(self, data: dict,config_object:dict=config.config_object,header:dict=config.header):
-        self.connector = connector.AdobeRequest(
-            config_object=config_object, header=header)
-        self.header = self.connector.header
-        self.endpoint = config.endpoints['global']
-        self.id = data['id']
-        self.name = data['attributes']['name']
-        self.state = data['attributes']['state']
-        self.build_required = data['attributes']['build_required']
-        self.builds = data['relationships']['builds']['links']['related']
-        self._DataElements = config.endpoints['global'] + \
-            '/libraries/'+data['id']+'/data_elements'
-        self._Extensions = config.endpoints['global'] + \
-            '/libraries/'+data['id']+'/extensions'
-        self._Environment = config.endpoints['global'] + \
-            '/libraries/'+data['id']+'/envrionment'
-        self._Rules = config.endpoints['global'] + \
-            '/libraries/'+data['id']+'/rules'
-        self._Builds = config.endpoints['global'] + \
-            '/libraries/'+data['id']+'/builds'
-        self.build_status = data['meta']['build_status']
-        self.relationships = {}
-        self._environments = {}
-        self._dev_env = ''
-
-    def getDataElements(self)->list:
-        """
-        retrieve the list of Data Elements attached to this library
-        """
-        dataElements = self.connector.getData(self._DataElements)
-        data = dataElements
-        # assign the list to its dict value
-        self.relationships['data_elements'] = data
-        return data
-
-    def getExtensions(self)->list:
-        """
-        retrieve the list of Extensions attached to this library
-        """
-        extensions = self.connector.getData(self._Extensions)
-        data = extensions
-        self.relationships['extensions'] = data
-        return data
-
-    def getRules(self)->list:
-        """
-        retrieve the list of rules attached to this library
-        """
-        rules = self.connector.getData(self._Rules)
-        data = rules
-        self.relationships['rules'] = data
-        return data
-
-    def getFullLibrary(self)->dict:
-        self.getDataElements()
-        self.getRules()
-        self.getExtensions()
-        return self.relationships
-
-    def addDataElements(self, data_element_ids: list)->dict:
-        """
-        Take a list of data elements id and attach them to the library. 
-        Arguments:
-            data_element_ids: REQUIRED : list of data elements id
-        """
-        if self.state != 'development':
-            print('State is not development, cannot add relationships')
-            return None
-        obj = {'data': []}
-        if type(data_element_ids) == str:
-            data_element_ids = data_element_ids.split(' ')
-        for ids in data_element_ids:
-            obj['data'].append({"id": ids,
-            "type": "data_elements", "meta": {"action": "revise"}})
-        url =  f'/libraries/{self.id}/relationships/data_elements'
-        res = self.connector.postData(self.endpoint +url, data=obj)
-        return res
-    
-    def updateDataElement(self,data_element_ids:list)->dict:
-        """
-        Update the data element inside the library. (PATCH)
-        Arguments:
-            data_element_ids: REQUIRED : list of data elements id
-        """
-        if self.state != 'development':
-            print('State is not development, cannot update relationships')
-            return None
-        obj = {'data': []}
-        if type(data_element_ids) == str:
-            data_element_ids = data_element_ids.split(' ')
-        for ids in data_element_ids:
-            obj['data'].append({"id": ids,
-            "type": "data_elements", "meta": {"action": "revise"}})
-        url =  f'/libraries/{self.id}/relationships/data_elements'
-        res = self.connector.patchData(self.endpoint +url, data=obj)
-        return res
-
-
-    def addRules(self, rules_ids: list)->dict:
-        """
-        Take a list of rules id and attach them to the library. 
-        Arguments:
-            rules_ids: REQUIRED : list of rules id
-        """
-        if self.state != 'development':
-            print('State is not development, cannot add relationships')
-            return None
-        obj = {'data': []}
-        if type(rules_ids) == str:
-            rules_ids = rules_ids.split(' ')
-        for ids in rules_ids:
-            obj['data'].append({"id": ids, "type": "rules",
-                                "meta": {"action": "revise"}})
-        url = f'/libraries/{self.id}/relationships/rules'
-        res = self.connector.postData(self.endpoint + url, data=obj)
-        return res
-    
-    def updateRules(self,rules_ids:list)->dict:
-        """
-        Replace all existing rules with the ones posted.
-        Arguments:
-            rules_ids: REQUIRED : list of rules id
-        """
-        if self.state != 'development':
-            print('State is not development, cannot add relationships')
-            return None
-        obj = {'data': []}
-        if type(rules_ids) == str:
-            rules_ids = rules_ids.split(' ')
-        for ids in rules_ids:
-            obj['data'].append({"id": ids, "type": "rules",
-                                "meta": {"action": "revise"}})
-        url = f'/libraries/{self.id}/relationships/rules'
-        res = self.connector.patchData(self.endpoint + url, data=obj)
-        return res
-
-    def addExtensions(self, extensions_ids: list)->object:
-        """
-        Take a list of extension id and attach them to the library. 
-        Arguments:
-            extensions_ids: REQUIRED : list of extension id
-        """
-        if self.state != 'development':
-            print('State is not development, cannot add relationships')
-            return None
-        obj = {'data': []}
-        if type(extensions_ids) == str:
-            extensions_ids = extensions_ids.split(' ')
-        for ids in extensions_ids:
-            obj['data'].append({"id": ids, "type": "extensions", 
-                                "meta": {"action": "revise"}})
-        url = f'/libraries/{self.id}/relationships/extensions'
-        res = self.connector.postData(self.endpoint+url, data=obj)
-        return res
-    
-    def updateExtensions(self, extensions_ids: list)->object:
-        """
-        Replace all existing extensions into the library. 
-        Arguments:
-            extensions_ids: REQUIRED : list of extension id
-        """
-        if self.state != 'development':
-            print('State is not development, cannot add relationships')
-            return None
-        obj = {'data': []}
-        if type(extensions_ids) == str:
-            extensions_ids = extensions_ids.split(' ')
-        for ids in extensions_ids:
-            obj['data'].append({"id": ids, "type": "extensions", 
-                                "meta": {"action": "revise"}})
-        url = f'/libraries/{self.id}/relationships/extensions'
-        res = self.connector.patchData(self.endpoint+url, data=obj)
-        return res
-
-    def setEnvironments(self, environments_list: list, dev_name: str = None)->None:
-        """
-        Save the different environments ids available. 
-        It is required to use the library class.
-        Arguments : 
-            environments_list : REQUIRED : list of environment retrieved by the getEnvironment method
-            dev_name : OPTIONAL : Name of your dev environment. If not defined, will take the first dev environment.
-        """
-        for env in environments_list:
-            if env['attributes']['stage'] == 'production':
-                self._environments['production'] = env['id']
-            elif env['attributes']['stage'] == 'staging':
-                self._environments['staging'] = env['id']
-            elif env['attributes']['stage'] == 'development':
-                devs = self._environments.get('developments', {})
-                devs[env['attributes']['name']] = env['id']
-                self._environments['developments'] = devs
-        if dev_name != None:
-            self._dev_env = self._environments['developments'][dev_name]
-        else:
-            key1 = list(self._environments['developments'].keys())[0]
-            self._dev_env = self._environments['developments'][key1]
-
-    def _setEnvironment(self, obj: dict,verbose:bool=False)->None:
-        path = f'/libraries/{self.id}/relationships/environment'
-        new_env = self.connector.patchData(self.endpoint+path, data=obj,verbose=verbose)
-        res = new_env
-        return res
-
-    def _removeEnvironment(self)->None:
-        """
-        Remove environment
-        """
-        path = f'/libraries/{self.id}/relationships/environment'
-        new_env = self.connector.getData(self.endpoint+path) 
-        return new_env
-
-    def build(self,verbose:bool=False)->dict:
-        """
-        Build the library. 
-        Part of the code takes care of assigning the right environement before building the library.
-        Returns the build when it is completed (succeed or not).
-        It will check every 15 seconds for the build status, making sure it is not "pending".
-        """
-        if self.build_required == False and self.state != 'approved':
-            return 'build is not required'
-        status = ""
-        if self.state == 'development':
-            env_id = self._dev_env
-            obj = {
-                "data": {
-                    "id": env_id,
-                    "type": "environments"
-                }
-            }
-            self._removeEnvironment()
-            status = self._setEnvironment(obj,verbose=verbose)
-        elif self.state == 'submitted':
-            env = 'staging'
-            obj = {
-                "data": {
-                    "id": self._environments[env],
-                    "type": "environments"
-                }
-            }
-            self._removeEnvironment()
-            status = self._setEnvironment(obj,verbose=verbose)
-        elif self.state == 'approved':
-            env = 'production'
-            obj = {
-                "data": {
-                    "id": self._environments[env],
-                    "type": "environments"
-                }
-            }
-            self._removeEnvironment()
-            status = self._setEnvironment(obj,verbose=verbose)
-        if 'error' in status.keys():
-            raise SystemExit('Issue setting environment')
-        build = self.connector.postData(self._Builds)
-        build_id = build['data']['id']
-        build_status = build['data']['attributes']['status']
-        while build_status == 'pending':
-            print('pending...')
-            time.sleep(20)
-            # return the json directly
-            build = self.connector.getData(
-                config.endpoints['global']+'/builds/'+str(build_id))
-            build_status = build['data']['attributes']['status']
-        if build['data']['attributes']['status'] == 'succeeded':
-            self.build_required = False
-            self.build_status = 'succeeded'
-        else:
-            self.build_required = True
-            self.build_status = build['data']['attributes']['status']
-        return build
-
-    def transition(self, action: str = None, **kwargs)->object:
-        """
-        Move the library along the publishing funnel.
-        If no action is provided, it would automatically go to the next state. 
-        Arguments : 
-            action : OPTIONAL : action to do on the library. Possible values: 
-                - 'submit' : if state == development
-                - 'approve' : if state == submitted
-                - 'reject' : if state == submitted
-        """
-        path = f'/libraries/{self.id}'
-        if action == None:
-            if self.state == 'development':
-                action = 'submit'
-            elif self.state == 'submitted':
-                action = 'approve'
-        obj = {"data": {
-            "id": self.id,
-            "type": "libraries",
-            "meta": {
-                "action": action
-            }
-        }
-        }
-        transition = self.connector.patchData(self.endpoint+path, data=obj)
-        data = transition
-        self.state = data['data']['attributes']['state']
-        self.build_required = data['data']['attributes']['build_required']
-        return data