+++ docs/releases.md
@@ -3,6 +3,15 @@
 This page gathered the changes made between version of the launchpy module.\
 This has been started after the 0.3.0 release.
 
+## 0.3.4
+
+* Adding a `getRuleComponent` method to retrieve singleComponent
+* Adding a `updateCustomCode` method to upload custom code directly to Launch
+* Adding a `updateProperty` method to update the property settings
+* separate `Library` class to its own submodule
+* cleaning code in launchpy core library
+* small fix `createConfigFile` method.
+
 ## 0.3.3
 
 * fixing the build method for library that didn't account for new optimized call with new architecture.